Account for power-cord disconection when sleeping
The last match of an AC log might not always reflect the correct
charge level as it might have not logged entries before the power
was disconected, for example, when disconecting the power cord
when the macbook was sleeping. Thus, it is necessary to find the
next BATT entry in the logs which will tell you exactly what the
charge was when the laptop was unplugged.